Rating to flaot

+++ schema/search_schema.py
@@ -15,7 +15,7 @@ class PreviewBranch(BaseModel):
     name: str
     image_url: Optional[str] = None  # TODO: Definir si es oopcional
     price: Optional[int] = 0  # TODO: Descartar Optional, por ahora se deja por poca e inconsistente data
-    rating: Optional[int] = 0
+    rating: Optional[float] = 0
     description: Optional[str] = None  # TODO: Creo que es opcional, por confirmar
